managment works

+++ BFH.EADN.Common/Types/Contracts/Management/Quiz.cs
@@ -14,9 +14,9 @@ namespace BFH.EADN.Common.Types.Contracts
         [DataMember(Order = 0)]
         public string Text { get; set; }
         [DataMember(Order = 0)]
-        public ushort MinQuestionCount { get; set; }
+        public int MinQuestionCount { get; set; }
         [DataMember(Order = 0)]
-        public ushort MaxQuestionCount { get; set; }
+        public int MaxQuestionCount { get; set; }
         [DataMember(Order = 0)]
         public QuizType Type { get; set; }
         [DataMember(Order = 0)]
