+++ BFH.EADN.Persistence.EF/Entities/Quiz.cs
@@ -15,8 +15,8 @@ namespace BFH.EADN.Persistence.EF.Entities
         public string Text { get; set; }
         public QuizType Type { get; set; }
         public DateTime? LastUsed { get; set; }
-        public ushort MinQuestionCount { get; set; }
-        public ushort MaxQuestionCount { get; set; }
+        public int MinQuestionCount { get; set; }
+        public int MaxQuestionCount { get; set; }
         public virtual ICollection<Question> Questions { get; set; }
     }
 }
